Restructuring Discord related content
