+++ README.md
@@ -1,1 +1,9 @@
 # WWW
+
+
+## Environment
+
+| Name               | Description                                           | Default |
+|--------------------|-------------------------------------------------------|---------|
+| `DISCORD_GUILD_ID` | Fetch Discord related data from specified guild       | -       |
+| `DISCORD_TOKEN`    | Discord Bot Token used to fetch data from Discord API | -       |
